Centroid an mzML file with configurable parameters

This function processes an mzML file for centroiding, providing detailed controls for peak picking, smoothing, and noise estimation. It allows fine-tuning of MS1 and MS2 peak detection, optimizing spectral data analysis for different experimental needs.

Usage

centroid_one_file(
  file,
  pattern,
  replacement,
  ms1_min_peaks = 1000,
  ms1_noise_estimator = c("SuperSmoother", "MAD"),
  ms1_peak_snr = 0,
  ms1_refine_mz = c("descendPeak", "kNeighbors", "none"),
  ms1_signal_percentage = 33,
  ms1_smooth_window = 6L,
  ms2_noise_estimator = c("SuperSmoother", "MAD"),
  ms2_peak_snr = 0,
  ms2_refine_mz = c("descendPeak", "kNeighbors", "none"),
  ms2_signal_percentage = 50,
  ms2_smooth_window = 4L
)

Arguments

file

character(1) Path to the input mzML file. The file must be accessible and in valid mzML format.

pattern

character(1) Regular expression pattern to match in the input file path, useful for modifying the output file path.

replacement

character(1) Replacement string for altering the output file path based on the pattern.

ms1_min_peaks

integer(1) Minimum number of peaks required to retain an MS1 spectrum. Spectra with fewer peaks than this threshold are discarded.

Default: 1000.

ms1_noise_estimator

character(1) Noise estimation method for MS1 spectra. Available options:

"SuperSmoother": Smoothing method from the supsmu function.

"MAD": Median Absolute Deviation

Default: "SuperSmoother".

ms1_peak_snr

numeric(1) Signal-to-noise ratio threshold for MS1 peak picking. Peaks below this threshold will be ignored. Default: 0.

ms1_refine_mz

character(1) Method for refining m/z values after MS1 peak picking. Available options:

"descendPeak": Refines peaks by descending intensity.

"kNeighbors": Refines using nearest neighbors.

"none": No refinement applied.

Default: "descendPeak".

ms1_signal_percentage

numeric(1) Minimum signal percentage (relative to the maximum signal) for retaining MS1 peaks in centroid calculation. Default: 33.

ms1_smooth_window

integer(1) Half window size for smoothing MS1 spectra. Larger values apply stronger smoothing.

Default: 6L.

ms2_noise_estimator

character(1) Noise estimation method for MS2 spectra. Options:

"SuperSmoother": Smoothing-based estimator.

"MAD": Median Absolute Deviation

Default: "SuperSmoother".

ms2_peak_snr

numeric(1) Signal-to-noise ratio threshold for MS2 peak picking. Default: 0.

ms2_refine_mz

character(1) Method for refining m/z values after MS2 peak picking. Available options:

"descendPeak": Refines peaks by descending intensity.

"kNeighbors": Refines using nearest neighbors.

"none": No refinement applied.

Default: "descendPeak".

ms2_signal_percentage

numeric(1) Minimum signal percentage (relative to the maximum signal) for retaining MS2 peaks in centroid calculation. Default: 50.

ms2_smooth_window

integer(1) Half window size for smoothing MS2 spectra. Larger values apply stronger smoothing. Default: 4L.

Value

logical(1) Returns TRUE if centroiding was successful; otherwise, returns FALSE.

Details

This function processes both MS1 and MS2 data, applying user-defined smoothing, peak-picking, and noise estimation methods. The centroiding is tailored according to specific experimental requirements, with file path customization available via pattern and replacement.

Author

Johannes Rainer, Adriano Rutz