Centroid an mzML file with configurable parameters

This function processes an mzML file for centroiding, providing detailed controls for peak picking, smoothing, and noise estimation. It allows fine-tuning of MS1 and MS2 peak detection, optimizing spectral data analysis for different experimental needs.

Usage

centroid_one_file(
  file,
  pattern,
  replacement,
  mz_tol_da_ms1 = 0.002,
  mz_tol_da_ms2 = 0.005,
  mz_tol_ppm_ms1 = 5,
  mz_tol_ppm_ms2 = 10,
  mz_fun = base::mean,
  int_fun = base::sum,
  mz_weighted = TRUE
)

Arguments

file

character(1) Path to the input mzML file. The file must be accessible and in valid mzML format.

pattern

character(1) Regular expression pattern to match in the input file path, useful for modifying the output file path.

replacement

character(1) Replacement string for altering the output file path based on the pattern.

mz_tol_da_ms1

numeric(1) m/z tolerance in Dalton (ms1).

Default: 0.002.

mz_tol_da_ms2

numeric(1) m/z tolerance in Dalton (ms2).

Default: 0.005.

mz_tol_ppm_ms1

numeric(1) m/z tolerance in ppm (ms1).

Default: 5.

mz_tol_ppm_ms2

numeric(1) m/z tolerance in ppm (ms2).

Default: 10.

mz_fun

Function to aggregate m/z values for all mass peaks within each peak group into a single m/z value. This parameter is ignored if weighted = TRUE (the default).

Default: base::mean.

int_fun

Function to aggregate intensities for all peaks in each peak group into a single intensity value.

Default: base::sum.

mz_weighted

logical(1) Boolean whether m/z values of peaks within each peak group should be aggregated into a single m/z value using an intensity-weighted mean.

Default: TRUE.

Value

logical(1) Returns TRUE if centroiding was successful; otherwise, returns FALSE.

Details

This function processes both MS1 and MS2 data, applying user-defined smoothing, peak-picking, and noise estimation methods. The centroiding is tailored according to specific experimental requirements, with file path customization available via pattern and replacement.

Author

Johannes Rainer, Adriano Rutz